Refactor : station관련 control 분리
<description>
station의 데이터를 관리하는 컨트롤러를 station.control.js파일에 넣었다.
