+++ src/View/input.js
@@ -1,13 +1,12 @@
 import {
   onChangeScreen,
-  onAddStation,
-  onRemoveStation,
   onAddLine,
   onRemoveLine,
   onLoadSection,
   onAddSection,
   onRemoveSection,
 } from '../index.js';
+import {onAddStation, onRemoveStation} from '../Controller/station-control.js';
 
 export const $screenButton = document.body.querySelectorAll('#app > button');
 export const $stationContainer = document.body.querySelector('#station');
